Ensure test filenames match test suite names and each file has a single suite defined
