+++ Tests/IntegrationTests/Containers/TestCLICopyCommand.swift
