+++ Tests/IntegrationTests/Containers/TestCLIStatsCommand.swift
